+++ relay-event-normalization/src/normalize/mod.rs
@@ -2265,6 +2265,7 @@ mod tests {
                 sentry_tags: ~,
                 received: ~,
                 measurements: ~,
+                _metrics_summary: ~,
                 other: {},
             },
         ]
@@ -2307,6 +2308,7 @@ mod tests {
                 sentry_tags: ~,
                 received: ~,
                 measurements: ~,
+                _metrics_summary: ~,
                 other: {},
             },
         ]
@@ -2349,6 +2351,7 @@ mod tests {
                 sentry_tags: ~,
                 received: ~,
                 measurements: ~,
+                _metrics_summary: ~,
                 other: {},
             },
         ]