+++ relay-server/src/actors/project.rs
@@ -1133,6 +1133,8 @@ impl Project {
         };
         let extraction_mode = ExtractionMode::from_usage(usage);
 
+        let enable_cardinality_limiter = project_state.has_feature(Feature::CardinalityLimiter);
+
         if !buckets.is_empty() {
             envelope_processor.send(EncodeMetrics {
                 buckets,
@@ -1140,6 +1142,7 @@ impl Project {
                 extraction_mode,
                 project_state,
                 rate_limits: self.rate_limits.clone(),
+                enable_cardinality_limiter,
             });
         }
     }